+++ docs/user/changes.rst
@@ -4,6 +4,8 @@ Release history
 - feat: new API :attr:`api.telephony.hotdesking_voiceportal <wxc_sdk.telephony.TelephonyApi.hotdesking_voiceportal>`
 - feat: new endpoint: :meth:`api.telephony.location.single_number_reach_primary_available_phone_numbers <wxc_sdk.telephony.location.TelephonyLocationApi.single_number_reach_primary_available_phone_numbers>`
 - feat: new API :attr:`api.telephony.cx_essentials.wrapup_reasons <wxc_sdk.telephony.cx_essentials.CustomerExperienceEssentialsApi.wrapup_reasons>`
+- feat: new API :attr:`api.person_settings.single_number_reach <wxc_sdk.person_settings.PersonSettingsApi.single_number_reach>`
+
 1.25.0
 ------
 - feat: new attributes :attr:`PSTNConnectionOption.route_type <wxc_sdk.telephony.pstn.PSTNConnectionOption.route_type>` and :attr:`.route_id <wxc_sdk.telephony.pstn.PSTNConnectionOption.route_type>`